editoast: move track_ranget to editoast_schemas
