+++ editoast/editoast_schemas/src/infra.rs
@@ -1,9 +1,11 @@
 mod direction;
 mod track_offset;
+mod track_range;
 mod waypoint;
 
 pub use direction::Direction;
 pub use track_offset::TrackOffset;
+pub use track_range::TrackRange;
 pub use waypoint::Waypoint;
 
 editoast_common::schemas! {
