+++ editoast/src/views/pathfinding/mod.rs
@@ -52,11 +52,11 @@ use crate::modelsv2::Retrieve as RetrieveV2;
 use crate::modelsv2::RollingStockModel;
 use crate::schema::ApplicableDirectionsTrackRange;
 use crate::schema::OperationalPoint;
-use crate::schema::TrackRange;
 use crate::schema::TrackSection;
 use crate::DbPool;
 use editoast_common::geometry::diesel_linestring_to_geojson;
 use editoast_common::geometry::geojson_to_diesel_linestring;
+use editoast_schemas::infra::TrackRange;
 
 crate::routes! {
     "/pathfinding" => {
